Alright fkn huge update XD
So in regards to

Registartion
Login
Profile page
Edit profile
Create group
Group Page
Edit group
Invite to group

Everything seems to be working correctly, it is communicating with the server and storing everything there. It is a hybrid setup (reason for so many different repository files) so it is also fetching some things from the server and then storing locally for caching purposes

in regards to

Events

I've gotten it working so that it is storing things on the server but the date time thing is just a placeholder as before and mostly just sending null to the server. Events are currently not showing on the group or user pages but that's most likely due to XML things.

I've done what I can but due to time constraints I need to work on other things for now.

Test out the branch and we'll be in touch

Ps. I also had to update some versions of plugins for android studio ask me about it if you have any issues.

+++ app/src/main/java/is/hi/hbv601_team_12/data/entities/EventParticipant.kt
@@ -29,8 +29,8 @@ import androidx.room.Index
     ]
 )
 data class EventParticipant(
-    val eventId: Int,
-    val userId: Int,
+    val eventId: Long,
+    val userId: Long,
     val status: ParticipantStatus = ParticipantStatus.INVITED,
 )
