+++ app/src/main/java/is/hi/hbv601_team_12/data/repositories/EventsRepository.kt
@@ -5,7 +5,7 @@ import kotlinx.coroutines.flow.Flow
 
 interface EventsRepository {
     fun getAllEventsStream(): Flow<List<Event>>
-    fun getEventStream(id: Int): Flow<Event>
+    fun getEventStream(id: Long): Flow<Event>  // Changed from Int to Long
     suspend fun insertEvent(event: Event)
     suspend fun updateEvent(event: Event)
     suspend fun deleteEvent(event: Event)
